image_processing.py
import cv2
import numpy as np

class RooftopAnalyzer:
    def __init__(self):
        self.pixels_per_meter = 50  # Adjust based on image resolution
        
    def analyze_image(self, image_path):
        img = cv2.imread(image_path)
        gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
        edges = cv2.Canny(gray, 50, 150)
        contours, _ = cv2.findContours(edges, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
        
        # Simple rooftop detection (replace with ML model in production)
        main_contour = max(contours, key=cv2.contourArea)
        area_px = cv2.contourArea(main_contour)
        area_m2 = area_px / (self.pixels_per_meter**2)
        
        return {
            'usable_area': area_m2 * 0.7,  # 70% usable assumption
            'orientation': 'South',       # Simplified for demo
            'shading_factor': 0.85
        }