requirements.txt
langchain-core
langchain-groq
streamlit
python-dotenv
opencv-python
numpy